Add some apis and fix some issues

+++ src/main/kotlin/moe/kurenai/tdlight/request/message/DeleteMessage.kt
@@ -8,7 +8,7 @@ import moe.kurenai.tdlight.request.Request
 
 data class DeleteMessage(
     val chatId: String,
-    val messageId: Long,
+    val messageId: Int,
 ) : Request<ResponseWrapper<Boolean>>() {
 
     @JsonIgnore
