+++ src/main/kotlin/moe/kurenai/tdlight/request/message/EditMessageMedia.kt
@@ -11,7 +11,7 @@ import moe.kurenai.tdlight.request.Request
 
 data class EditMessageMedia(
     val media: InputFile,
-    val disableWebPagePreview: Boolean? = null, override val chatId: String? = null, override val messageId: Long? = null, override val inlineMessageId: String? = null,
+    val disableWebPagePreview: Boolean? = null, override val chatId: String? = null, override val messageId: Int? = null, override val inlineMessageId: String? = null,
 ) : Request<ResponseWrapper<Message>>(), EditMessageRequest, MediaRequest {
 
     @JsonIgnore
